requirements.txt
amqp==5.3.1
asgiref==3.11.1
async-timeout==5.0.1
attrs==26.1.0
autobahn==24.4.2
Automat==25.4.16
billiard==4.2.4
celery==5.6.3
cffi==2.0.0
channels==4.1.0
channels_redis==4.2.1
click==8.3.3
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==47.0.0
daphne==4.1.2
dj-database-url==3.1.2
Django==5.2.0
django-cors-headers==4.9.0
django-filter==25.2
django-redis==6.0.0
django_celery_results==2.6.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
exceptiongroup==1.3.1
gunicorn==25.3.0
hyperlink==21.0.0
idna==3.13
Incremental==24.11.0
inflection==0.5.1
iniconfig==2.3.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kombu==5.6.2
msgpack==1.1.2
packaging==26.2
pillow==12.2.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.12
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
PyJWT==2.12.1
pyOpenSSL==26.1.0
pytest==8.1.1
pytest-asyncio==0.23.6
pytest-django==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
redis==7.4.0
referencing==0.37.0
rpds-py==0.30.0
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.5
tomli==2.4.1
Twisted==25.5.0
txaio==25.9.2
typing_extensions==4.15.0
tzdata==2026.2
tzlocal==5.3.1
uritemplate==4.2.0
vine==5.1.0
wcwidth==0.6.0
whitenoise==6.12.0
zope.interface==8.4
